Django/3일차과제/pyproject.toml
[tool.poetry]
name = "django-study"
version = "0.1.0"
description = "python django project"
authors = ["kyoorochi <pmh0521@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
